Rename Shared/Tests.cpp to Tests.h and add it to UnitTests CMakeLists
