+++ Apps/UnitTests/X11/App.cpp
@@ -3,7 +3,7 @@
 #include <X11/Xlib.h> // will include X11 which #defines None... Don't mess with order of includes.
 #include <X11/Xutil.h>
 #undef None
-#include "../Shared/Tests.cpp"
+#include "../Shared/Tests.h"
 
 namespace
 {
